Move RustCrypto module into a separate file
